Updated unit tests on MovieDetail class methods, switched movieRating type to Double on Movie, MovieDetail, MovieDetailResponse, MovieListDB and MovieListResponse classes

+++ MovieDBPI/app/src/main/java/com/mariana/moviedbpi/data/model/response/MovieDetailResponse.kt
@@ -20,7 +20,7 @@ data class MovieDetailResponse(
     val releaseDate: String,
 
     @SerializedName("vote_average")
-    val userRating: Float,
+    val userRating: Double,
 
     @SerializedName("runtime")
     val runtime: Int,
